+++ CHANGELOG.md
@@ -17,6 +17,8 @@ CHANGELOG
 - Upgrade Slurm to 23.11.1.
 - Add support for Python 3.11, 3.12 in pcluster CLI and aws-parallelcluster-batch-cli.
 - Upgrade Python to version 3.12 and NodeJS to version 18 in ParallelCluster Lambda Layer.
+- Build network interfaces using network card index from `NetworkCardIndex` list of EC2 DescribeInstances response, 
+  instead of looping over `MaximumNetworkCards` range.  
 
 3.8.0
 ------